fix file preview'

+++ app/Http/Controllers/DokumenController.php
@@ -81,19 +81,21 @@ class DokumenController extends Controller
         $dokumen->file_list = '';
         foreach ($dokumen->files as $file) {
             $ext = pathinfo($file->file, PATHINFO_EXTENSION);
-            $src = exts(strtolower($ext));
-            $dokumen->dokumen_list .= '<div class="col-span-12 lg:col-span-4" id="dokumendiv'.$file->id.'" style="position:relative;">
-                <div style="height: 100px;">
-                    <img class="img-fluid card-img-top" src="'.asset('images').'/'.$src.'" alt="File'.$file->id.'" style="max-height: 100px; max-width:100%; padding: 5px 0;">
-                </div>
-                <div class="form-group mb-0">
-                    <input type="text" name="deskripsi_existing['.$file->id.']" class="form-control filenya" id="deskripsi'.$file->id.'" placeholder="Deskripsi" value="'.$file->deskripsi.'">
-                    <input type="hidden" name="file_existing['.$file->id.']" value="'.$file->file.'">
-                </div>
-                <a href="javascript:void(0);" onclick="removeFile('.$file->id.')" class="remove-button text-danger">
-                    <div class="tooltip w-5 h-5 flex items-center justify-center absolute rounded-full text-white bg-danger right-0 top-0 -mr-2 -mt-2"> <svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round" icon-name="x" data-lucide="x" class="lucide lucide-x w-4 h-4"><line x1="18" y1="6" x2="6" y2="18"></line><line x1="6" y1="6" x2="18" y2="18"></line></svg> </div>
-                </a>
-            </div>';
+            if (in_array($ext, exts())) {
+                $src = exts(strtolower($ext));
+                $dokumen->dokumen_list .= '<div class="col-span-12 lg:col-span-4" id="dokumendiv'.$file->id.'" style="position:relative;">
+                    <div style="height: 100px;">
+                        <img class="img-fluid card-img-top" src="'.asset('images').'/'.$src.'" alt="File'.$file->id.'" style="max-height: 100px; max-width:100%; padding: 5px 0;">
+                    </div>
+                    <div class="form-group mb-0">
+                        <input type="text" name="deskripsi_existing['.$file->id.']" class="form-control filenya" id="deskripsi'.$file->id.'" placeholder="Deskripsi" value="'.$file->deskripsi.'">
+                        <input type="hidden" name="file_existing['.$file->id.']" value="'.$file->file.'">
+                    </div>
+                    <a href="javascript:void(0);" onclick="removeFile('.$file->id.')" class="remove-button text-danger">
+                        <div class="tooltip w-5 h-5 flex items-center justify-center absolute rounded-full text-white bg-danger right-0 top-0 -mr-2 -mt-2"> <svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round" icon-name="x" data-lucide="x" class="lucide lucide-x w-4 h-4"><line x1="18" y1="6" x2="6" y2="18"></line><line x1="6" y1="6" x2="18" y2="18"></line></svg> </div>
+                    </a>
+                </div>';
+            }
         }
         return response()->json($dokumen);
     }
